* ojojoj, didn't know its fuxed 0.o
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/trunk@1196 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ configmanager.h
@@ -212,7 +212,7 @@ class ConfigManager
 		std::string getGlobalString(lua_State* _L, const std::string& _identifier, const std::string& _default = "");
 		bool getGlobalBool(lua_State* _L, const std::string& _identifier, const std::string& _default = "no");
 		int32_t getGlobalNumber(lua_State* _L, const std::string& _identifier, const int32_t _default = 0);
-		double getGlobalDouble(lua_State* _L, const std::string& _identifier, const int32_t _default = 0);
+		double getGlobalDouble(lua_State* _L, const std::string& _identifier, const double _default = 0);
 
 		bool m_loaded, m_startup;
 		lua_State* L;
